refatorando o componente PostCard
